Create uwtools.api.execute
